refactor(features): 配置 `features/commands/`
将包含注册命令的模块移入
