+++ src/features/commands/handlers/showCommandsList.ts
@@ -1,4 +1,4 @@
-import { commandManager } from "../core/command";
+import { commandManager } from "../../../core/command";
 
 commandManager.registerCommand(['showshowway', '假人命令列表'], ({ entity }) => {
     entity?.sendMessage(commandManager.listRegisteredPrefixes().join('\n'));
