+++ src/features/commands/triggers/chat-send.ts
@@ -1,6 +1,6 @@
 import { system, world } from "@minecraft/server";
-import { commandManager, getLocationFromEntityLike, CommandError } from "../core/command";
-import { Messages } from "../constants/messages";
+import { commandManager, getLocationFromEntityLike, CommandError } from "../../../core/command";
+import { Messages } from "../../../constants/messages";
 
 world.beforeEvents.chatSend.subscribe(({message, sender}) => {
     system.run(() => {
