+++ src/features/index.ts
@@ -1,11 +1,11 @@
-export * from './help'
-export * from './chatSpawn'
-export * from './behaviors'
-export * from './breakBlock'
-export * from './youAreMine'
+export * from './commands/handlers/help'
+export * from './commands/handlers/chatSpawn'
+export * from './commands/handlers/behaviors'
+export * from './commands/handlers/breakBlock'
+export * from './commands/handlers/youAreMine'
 export * from './task'
 export * from './gui'
 export * from './killedBySimPlayer'
-export * from './setting'
-export * from './showCommandsList'
+export * from './commands/handlers/setting'
+export * from './commands/handlers/showCommandsList'
 export * from './tps/indicator'
