+++ src/main.ts
@@ -3,7 +3,7 @@ import { world } from '@minecraft/server'
 import './features'
 
 import {playerReady} from "./features/events/player-ready";
-import './triggers'
+import './features/commands/triggers'
 import { SimulatedPlayerManager } from './core/simulated-player';
 import { GameTestManager } from './core/gametest/manager';
